requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
pydantic-settings>=2.0.0
slowapi>=0.1.9
boto3>=1.34.0
mangum>=0.17.0
